Problem_scrapping_agent/pyproject.toml
[project]
name = "problem-scrapping-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "praw",
    "google-generativeai",
    "crewai",
    "beautifulsoup4",
    "python-dotenv"
]
